Make application home page.
Home page contain with 'Title', 'Sub-title', 'Button'.

*TODO: Make food page contains only food text.
