+++ src/App.jsx
@@ -2,6 +2,8 @@ import React from 'react';
 
 import styled from '@emotion/styled';
 
+import HomePage from './HomePage';
+
 const GridLayout = styled.div({
   display: 'grid',
   gridTemplateColumns: '1fr minmax(480px, 1fr) 1fr',
@@ -45,10 +47,7 @@ export default function App() {
             <h2>오늘 뭐 먹지?!</h2>
           </AppHeader>
           <AppContent>
-            <h1>오늘 뭐 먹지?!</h1>
-            <p>무엇을 먹을지 고민이라면,</p>
-            <p>오늘 한 끼 여기에 맡겨라!</p>
-            <button type="button">메뉴 뽑기</button>
+            <HomePage />
           </AppContent>
           <AppFooter />
         </AppContainer>
