+++ src/HomePage.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+export default function HomePage() {
+  return (
+    <>
+      <h1>오늘 뭐 먹지?!</h1>
+      <p>무엇을 먹을지 고민이라면,</p>
+      <p>오늘 한 끼 여기에 맡겨라!</p>
+      <button type="button">메뉴 뽑기</button>
+    </>
+  );
+}
